+++ src/mem.h
@@ -20,11 +20,14 @@
 extern "C" {
 #endif
 
+#include <stdbool.h>
+#include <stdint.h>
+
 void MemInit();
 void *xMemAlloc(size_t);
 void xMemFree(void *);
-int xMemGetUsed();
-int xMemGetSize(void *);
+int16_t xMemGetUsed();
+int16_t xMemGetSize(void *);
 void MemClear();
 
 #ifdef __cplusplus
